Stop game image from changing when disk removed
Size of game image was determined by the disks in play on each move in
the game. When the largest disk was removed, the game screen would
change size.

Now, game image size is determined at the start of the game and does not
change from there on in.

Also, added numbers to the disks to make gameplay easier.

Changes to be committed:
      modified:   hanoi.rkt