Composed the two components!
Changed the way that the components communicate with each other.

The component that composes the pieces is called the *controller
component*.
- Each component now has a `final-state` function, which takes a valid
state for the component and creates the state that other components
should expect to work with. For instance, the menu component is for
choosing a value from a selection of values. So, the final state for the
menu component is the chosen value. It should not be up to another
component to figure out what value was chosen, since the menu component
has all the necessary information to figure this out.
- The way that I composed each followed this rough pattern:
    - Changes from one component to another necessarily occur during
    functions that can produce a new state, such as on-key. So that's
    where the logic for transitioning from one component to another is.
    - The private state of the controller keeps track of which component
    is the one currently producing images for the user.
    - When producing the next state of the current component, we check
    if that component finished running its course by using the
    component's stop-when function (see lines 54 and 61 of hanoi.rkt for
    examples). If that component finished, then we use the component's
    final state function to create the state that the next component
    will consume.
    - Components take some initial arguments. The initial argument
    should be the final state of the previous component. Each component
    should be designed to work with the final state of the previous
    component. NOTE: This makes the idea of initial-state->state
    useless, for the most part. It's easier for the current component to
    produce a state that abides by some contract (final-state function)
    than it is for the next component to have the burden of dealing with
    any possible internal state of the current component
    (initial-state->state).

Problems:

- big-bang does not resize it's screen during runtime. The 1st image is
a scene that all other images are placed into.
- For components whose properties depend on the state of a previous
component (like towers requiring a tower-height to know what to do), I
have to rebuild the component on every render, which is no good. I think
I'll have to use non-functional functions to do this right. Like a
professor once told me, the involvement of time here is at least
limited, and the burden of thinking about it can be placed in somewhere
that the user doesn't have to think about. If I put the time-based
burdens in this composition function, then I don't have to design each
individual component with that burden. In other words, I can solve the
problem of time (switching components) just once, and reuse that
solution elsewhere.
- I haven't managed to generalize the controller component. Composition
should be an operation that is the same for all the components.
- This application necessarily requires you to build independent
components which feed into each other. Games don't really work like
that. What if I wanted a pause menu that came up while playing the
towers game, so that I could restart, or change the tower size, or just
get the solution? Plausible scenarios, and they all need a menu that
combines with the game, even though each piece can be designed in
isolation.

Changes to be committed:
      modified:   hanoi.rkt
      modified:   menu.rkt
      modified:   towers.rkt