Made composition a little more general
I've put the current component in the private state of the controller
component. With this, I've avoided using `set!`, but not sure if that
was really a good move. But I'd rather introduce the use of `set!` where
neccessary.

I've also added in an initial scene size, but it's a quick-and-dirty
hack, and it cannot possibly account for all scenarios.

Added names to each component as a temporary measure. Might keep them
around. Not sure.

I think next step is to maintain a list of components left to render,
and the current component. The controller component can end itself when
it sees that there are no components left to render.