Change initial-state->state to initial-state
The function form was useless, because that function was overtaken by
the initializer argument to functions like make-menu-game or
make-towers-game. Initialization happened there, and now each game had
to specify the initial state they wanted. That initial state is informed
by the initializer. So the initial-state is free to be a non-procedure
value now.

Changes to be committed:
      modified:   compose-components.rkt
      modified:   hanoi.rkt
      modified:   menu.rkt
      modified:   towers.rkt