Change final-state of components to output
Part of a component's interface was the expected result from it's
final-state function. it was confusing to explain that the input to this
function was the "true" final state of the component (the state that
caused the component's stop-when function to return true), and that the
final-state function took a "true" final state and returned the a value
that was meant for the consumption of other components.

Renaming this to output makes more sense. The output function consumes
the final state of a component and produces some value that other
components can depend on.

Changes to be committed:
      modified:   compose-components.rkt
      modified:   hanoi.rkt
      modified:   menu.rkt
      modified:   towers.rkt